+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Changelog
 
 ## 0.7.0
 - expect:body added support for regex matching when starting with `r`, defaults to `r".*<input>.*"`.
+- default port /metrics to 9080
 
 ## 0.6.0
 - Allow POST, PUT, DELETE, PATCH, OPTIONS, HEAD, TRACE, CONNECT methods.